enhanced: Keep a cumulative count of bytes read and written
The node exporter maintains this as a counter of the total bytes
read and written since system boot. RDS enhanced metrics do not
give us this information, instead presenting the number of bytes
read and written in each monitoring interval.

This means that the node exporter metrics we produce here were
previously incorrect. Instead of directly exposing the information
provided by the RDS metrics, maintain a counter to which we add the
additional bytes read/written on each scrape.